Create file.txt
creating a folder for the csv files that I used MAINLY for jupyter notebook. Many of the csv's wouldn't completely upload when to MYSql Workbench. 

In the instances the csv's wouldn't upload, it was usually one for lbj's season, I would go to a csv:sql conversion website and write the appropriate code in Workbench to create the SQL file.